refactor(baseball): change package struct
패키지 구조를 좀 더 세분화하는 방향으로 변경하였습니다.
