+++ src/main/java/baseball/utils/RandomNumberPicker.java
@@ -1,4 +1,4 @@
-package baseball.modules;
+package baseball.utils;
 
 import java.util.HashSet;
